Move Sx methods to a new HasSx class
