+++ src/main/java/org/sagebionetworks/web/client/jsinterop/SxProps.java
@@ -1,11 +1,12 @@
 package org.sagebionetworks.web.client.jsinterop;
 
+import elemental2.core.JsObject;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
 @JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "Object")
-public class SxProps {
+public class SxProps extends JsObject {
 
   public String color;
